server_python/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
opencv-python
numpy
pillow
